@@ -2,35 +2,39 @@
   "name": "thisisnotavirus",
   "version": "0.1.0",
   "private": true,
+  "type": "module",
   "dependencies": {
-    "@babel/core": "^7.24.7",
-    "@babel/preset-env": "^7.24.7",
-    "@babel/preset-typescript": "^7.24.7",
-    "@sentry/browser": "^8.28.0",
-    "@types/mustache": "^4.2.5",
-    "@types/three": "^0.152.1",
-    "@types/ua-parser-js": "^0.7.39",
-    "babel-loader": "^9.1.3",
-    "concurrently": "^8.2.2",
-    "css-loader": "^6.11.0",
-    "dotenv": "^16.4.5",
-    "eslint": "^8.57.0",
-    "mustache": "^4.2.0",
-    "prettier": "^2.8.8",
-    "sass": "^1.77.6",
-    "sass-loader": "^13.3.3",
-    "serve": "^14.2.3",
-    "style-loader": "^3.3.4",
+    "@babel/core": "7.24.7",
+    "@babel/preset-env": "7.24.7",
+    "@babel/preset-typescript": "7.24.7",
+    "@eslint/js": "9.27.0",
+    "@sentry/browser": "8.28.0",
+    "@types/mustache": "4.2.5",
+    "@types/three": "0.152.1",
+    "@types/ua-parser-js": "0.7.39",
+    "babel-loader": "9.1.3",
+    "concurrently": "8.2.2",
+    "css-loader": "6.11.0",
+    "dotenv": "16.4.5",
+    "eslint": "9.27.0",
+    "glob": "10.3.10",
+    "globals": "14.0.0",
+    "mustache": "4.2.0",
+    "prettier": "2.8.8",
+    "sass": "1.77.6",
+    "sass-loader": "13.3.3",
+    "serve": "14.2.3",
+    "style-loader": "3.3.4",
     "three": "0.153.0",
-    "ua-parser-js": "^1.0.38",
-    "webpack": "^5.92.1",
-    "webpack-cli": "^5.1.4"
+    "ua-parser-js": "1.0.38",
+    "webpack": "5.92.1",
+    "webpack-cli": "5.1.4"
   },
   "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "^7.15.0",
-    "@typescript-eslint/parser": "^7.15.0",
-    "source-map-loader": "^4.0.2",
-    "typescript": "^5.5.3"
+    "@typescript-eslint/eslint-plugin": "8.32.1",
+    "@typescript-eslint/parser": "8.32.1",
+    "source-map-loader": "4.0.2",
+    "typescript": "5.5.3"
   },
   "scripts": {
     "start": "concurrently 'yarn:build --watch' 'yarn:serve'",
